% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_displacement.R
\name{export_displacement}
\alias{export_displacement}
\title{export_displacement}
\usage{
export_displacement(pca_mods_object, magnification = NULL, dimx = 0,
  dimy = 0, dimz = 0)
}
\arguments{
\item{pca_mods_object}{list: output of the pca_mods function.}

\item{magnification}{numeric: magnification factor for shape variations.}

\item{dimx}{numeric: x voxel size.}

\item{dimy}{numeric: y voxel size.}

\item{dimz}{numeric: z voxel size.}

\item{interactive_plot}{logical: if TRUE the user can click on the PCA to visualize the associated shape.}
}
\description{
This function export the displacement file to be visualized into Paraview in an interactive way
}
\author{
Antonio Profico, Paul O'Higgins
}
