% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.fold.R
\name{mod.fold}
\alias{mod.fold}
\title{mod.fold}
\usage{
mod.fold(folder.path)
}
\arguments{
\item{folder.path}{path of the folder. If NULL use a Windows shell folder widget to choose a folder interactively}
}
\value{
position numeric: a vector of the row indices

data numeric: an array containing the model/displacement files exported from Paraview
}
\description{
This function stores into an array the model and displacement files from Paraview
}
\author{
Antonio Profico, Paul O'Higgins
}
