% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_shape_v.R
\name{pca_shape_v}
\alias{pca_shape_v}
\title{pca_shape_v}
\usage{
pca_shape_v(pca_mods_object, magnification = 100, interactive_plot = TRUE,
  PCx = 1, PCy = 2, PCXs = NULL, PCYs = NULL, PCscores = NULL)
}
\arguments{
\item{pca_mods_object}{list: output of the pca_mods function.}

\item{magnification}{numeric: magnification factor for shape variations.}

\item{interactive_plot}{logical: if TRUE the user can click on the PCA to visualize the associated shape.}

\item{PCx}{numeric: PCscore to be plotted on x axis.}

\item{PCy}{numeric: PCscore to be plotted on x axis.}

\item{PCXs}{numeric: insert the values of PCscore for the shape variation requested.}

\item{PCYs}{numeric: insert the values of PCscore for the shape variation requested.}
}
\description{
This function visualize shape variations using the output of the pca_mods function
}
\author{
Antonio Profico, Paul O'Higgins
}
