% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_shape_v.R
\name{tps_displ}
\alias{tps_displ}
\title{tps_displ}
\usage{
tps_displ(tar,ref,mag,sel)
}
\arguments{
\item{tar}{numeric: matrix k x 3, matrix to be warped.}

\item{ref}{numeric: matrix k x 3, reference matrix.}

\item{mag}{numeric: magnification factor.}

\item{sel}{numeric: subselection of voxels for tps warping.}
}
\description{
This function performs define a new matrix via TPS.
}
\author{
Antonio Profico, Paul O'Higgins
}
